\name{extract-methods}
\alias{extract-methods}
\alias{[.spectra.matrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
extract methods}
\description{
extract methods}
\usage{
\method{[}{spectra.matrix} (x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Object of class \code{spectra.matrix}.}
  \item{...}{
Additional args.}
}
\value{
Returns an object of class \code{spectra.matrix}.
}
\author{
Daniel M Griffith}

\examples{

#data(shootout)
#shootout_scans[1:10,]

}
\keyword{manipulation}
