% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idSpectra.R
\name{idSpectra}
\alias{idSpectra}
\title{Identify spectra in a plot.}
\usage{
idSpectra(spec)
}
\arguments{
\item{spec}{An object of class \code{spectra.list} or \code{spectra.matrix}.}
}
\value{
Returns an the name of the selected scan.
}
\description{
This function accepts spectra (the same ones given to \code{plot()}), allows
the user to click a scan, then returns the scan name in the console and
colors the selected scan red. This is useful is you are using the 
\code{plotSpectra()} with \code{base_plot == TRUE}, although the \code{plotly}
interactive version is preferable and already interactive.
}
\examples{


#data(shootout)
#plot(shootout_scans, base_plot == TRUE)
#idSpectra(shootout_scans)

}
\author{
Daniel M Griffith
}
\keyword{visualization}
