% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizePLS.R
\name{optimizePLS}
\alias{optimizePLS}
\title{Function to optimize the preprocessing steps for a PLS model.}
\usage{
optimizePLS(component, spectra, training_set = NULL, parallel = FALSE,
  region_list = NULL, preprocessing_list = NULL, max_comps = 10)
}
\arguments{
\item{component}{A vector of y-values. One for each spectrum.}

\item{spectra}{An object of class \code{spectra.matrix} containing spectra.
Rows should be in the same order as the component y-values.}

\item{training_set}{A logical vector of \code{length(component)} specifying
\code{TRUE} for training/calibration data and \code{FALSE} for
test/validation set data.}

\item{parallel}{Logical. The default is \code{FALSE}; \code{TRUE} allows for
the parallelization of validation proceedures, using the number of available
cores - 1. If \code{FALSE} the function will not be parallelized.}

\item{region_list}{A list, where each element is a vector of length 2,
specifying the range (max/min) of a spectral region to select. Should be in 
the same units as your spectra (e.g., wavenumbers).}

\item{preprocessing_list}{A list, where each element is either (1) a single
character string specifying a preprocessing step or (2) a vector of length 2
specifying a series of preprocessing steps to be applied together. See 
documentation for \code{preprocess()} for all available options.}

\item{max_comps}{What is the maximum number of latent vectors to possibly
include in the PLS regression. An integer.}
}
\value{
Returns an object of class \code{PLSopt}. The object is a list
containing the optimization results. See the following:

optimization_results - a data.frame containing the RMSEP for each
combination of preprocessing and subsetting tested\cr param_subsets - a list
of regions tried\cr param_preproc - a list of preprecessing steps tried.
}
\description{
This function iterates all combinations of preprocessing steps and spectral
region subsets, fitting PLS models for each, and ranks the resulting models
according to the Root Mean Squared Error of Prediction (or CV for cross
validation). This process informs the selection of parameters for fitting
PLS models with the function \code{calibrate()}. The function also
determines the rank (number of latent vectors) that is optimal for each
model.
}
\details{
By default the spectral regions are defined in wavenumbers as list(c(9400,7500), c(7500,6100), c(6100,5450), c(5450,4600), c(4600,4250)).
}
\examples{


# See main leaf.spec-package example.


}
\author{
Daniel M Griffith
}
\keyword{calibration}
