% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspec-package.R
\docType{package}
\name{plantspec-package}
\alias{plantspec-package}
\alias{plantspec}
\title{plantspec: NIR Calibration and Spectral Data Management in R}
\description{
This R package provides access to a global dataset of leaf NIR spectra and
contains tools for developing Partial Least Squares (PLS) regression models
for predicting leaf chemistry. Users can make use of pre-fitted
calibrations, add new data and fit new calibrations, and manage their
calibrations and datasets. The package represents an efficient workflow for
producing robust spectral calibrations.
}
\details{
Resources include:\cr - Functions for reading and writing spectra.\cr -
Tools for spectral manipulation and management (e.g., preprocessing,
conversion, subsetting).\cr - Tools for sample selection and experimental
design (e.g., Kennard-Stone selection).\cr - Wrapper functions for
optimization and fitting of spectral calibration models.\cr - Methods for
plotting spectra and models.\cr - Data and global calibrations for leaf C,
N, P, and K, documented in the companion data package
\code{leaf.spec.DB}.\cr - Data for examples, documented in \code{shootout}.

See the example below for a simple workflow for fitting a spectral
calibration model for nitrogen in a sample dataset.
}
\examples{


## LOAD TEST CALIBRATION DATA
#  data(shootout)
## INSPECT SPECTRA
#  plot(shootout_scans)
## CONVERT TO WAVENUMBERS
#  shootout_scans <- convertSpectra(x = shootout_scans, method = "WL_to_WN")
## SELECT TRAINING DATA FOR FITTING THE PLS MODEL
#  training_set <- !(subdivideDataset(spectra = shootout_scans, 
#                                     component = shootout_wetlab$N, 
#                                     method = "SPXY",
#                                     p=0.2, 
#                                     type = "validation"))
## OPTIMIZE SPECTRAL PREPROCESSING AND REGION SELECTION
#  N_opt <- optimizePLS(component = shootout_wetlab$N,
#                       spectra = shootout_scans,
#                       training_set = training_set)
## FIT A PLS MODEL, USING THE OPTIMAL PARAMETERS
#  N_cal_shootout <- calibrate(component=shootout_wetlab$N,
#                       spectra=shootout_scans,
#                       optimal_params=N_opt,
#                       optimal_model=1, 
#                       validation = "testset", 
#                       training_set = training_set)
## INSPECT CALIBRATION AND VALIDATION R2
#  N_cal_shootout$R2_Cal
#  N_cal_shootout$R2_Val

## PLOT FITTED vs TRUE VALUES 
#  plot(N_cal_shootout)

## SAVE MODEL AND OPTIMIZATION RESULTS FOR FUTURE USE (i.e., new data prediction)
# save(N_cal_shootout, Ncal, file = "N_calibration.RData")
## LOAD MODEL FOR USE IN PREDICTION OF NEW SPECTRA
# load("N_calibration.RData")


}
\references{
Bjørn-Helge Mevik, Ron Wehrens and Kristian Hovde Liland
(2013). pls: Partial Least Squares and Principal Component regression. R
package version 2.4-3. http://CRAN.R-project.org/package=pls
}
\author{
Daniel M Griffith <griffith.dan@gmail.com>; T. Michael Anderson
<anderstm@wfu.edu>
}
\keyword{package}
