% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPLScalibration.R
\name{plotPLScalibration}
\alias{plotPLScalibration}
\title{Visualize PLS modeling results.}
\usage{
plotPLScalibration(x, pch = 21, bg = "gray", ncomp = NULL,
  plottype = "prediction", ...)
}
\arguments{
\item{x}{An object of class \code{PLScalibration} containing the model.}

\item{pch}{see par.}

\item{bg}{see par.}

\item{ncomp}{The number of latent vectors to use. If NULL, the default is to
use the optimal number of determined during the optimization process.}

\item{plottype}{One of:\cr "prediction" - Predicted versus observed.\cr
"validation" - RMSEP versus rank.\cr "coefficients" - Coefficients.\cr
"scores" - Biplot of scores.\cr "loadings" - Loadings.}

\item{...}{Additional args.}
}
\value{
A plot of model results.
}
\description{
This function is a wrapper to the plot methods in the pls R package. It
takes an object of class \code{PLScalibration} and produces diagnostic
plots.
}
\examples{


#data(N_cal_shootout)
#plot(N_cal_shootout)

}
\author{
Daniel M Griffith
}
\keyword{visualization}
