% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Function to plot spectra.}
\usage{
plotSpectra(x, col = "black", type = NULL, xlim = NULL,
  ylim = NULL, add = FALSE, base_plot = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{spectra.list} or \code{spectra.matrix}
containing the spectra to plot.}

\item{col}{see par.}

\item{type}{see par.}

\item{xlim}{see par.}

\item{ylim}{see par.}

\item{add}{see par.}

\item{base_plot}{Default is FALSE. If FALSE, plots spectra as an interactive \code{plotly} object.}

\item{...}{Additional args.}
}
\description{
This function accepts spectra in a \code{spectra.list} or
\code{spectra.matrix} object and plots them.
}
\examples{


#data(shootout)
#plot(shootout_scans)

}
\author{
Daniel M Griffith
}
\keyword{visualization}
