\name{shootout}
\alias{shootout_scans}
\alias{shootout_wetlab}
\docType{data}
\title{
Sample spectral data: software shootout at the IDRC98
}
\description{
Sample NIR and wet lab data for illustration. Dry plant material (Fescue grass) from a fertilizer experiment were ground in a Wiley mill and scanned on a FOSS NIRSystems spectrophotometer in 1998. Carbon and nitrogen data were determined independently with a LECO CNS-2000.

For additional details about these data, visit:

\url{http://asrg.contactincontext.org/asrg/cnirs/Shoot\%20Out\%201998/shoot\%20out\%201998.html}

}
\usage{data(shootout)}
\format{
 shootout_scans - a \code{spectra.matrix} containing the sample NIR spectra.
 shootout_wetlab - a \code{data.frame} with the nitrogen and carbon data.
}

\references{
McClure, W., 1998. Software shootout at the idrc98. In: The Ninth International Diffuse Reflectance Conference, Chambersburg, Pennsylvania.
}
\examples{
#data(shootout)
}
\keyword{datasets}
