% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.spectra.list.R
\name{as.spectra.list}
\alias{as.spectra.list}
\title{Function to convert data to an object of class \code{spectra.list}.}
\usage{
as.spectra.list(x, ...)
}
\arguments{
\item{x}{Object of class \code{spectra.matrix}.}

\item{...}{Additional args.}
}
\value{
Returns an object of class \code{spectra.list}.
}
\description{
Convert data to an object of class \code{spectra.list} from an object of
class \code{spectra.matrix}.
}
\examples{


#data(shootout)
#scans <- as.spectra.list(shootout_scans)

}
\author{
Daniel M Griffith
}
\keyword{manipulation}
