% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSpectra.R
\name{combineSpectra}
\alias{combineSpectra}
\title{Function to combine spectral datasets into one object.}
\usage{
combineSpectra(x, y)
}
\arguments{
\item{x}{An object of class \code{spectra.list} or \code{spectra.matrix}.}

\item{y}{An object of class \code{spectra.list} or \code{spectra.matrix}.}
}
\value{
Returns a \code{spectra.matrix}.
}
\description{
This function accepts two objects of class \code{spectra.list} or
\code{spectra.matrix} and combines them into one \code{spectra.matrix}.
These objects should be in the same units and of the same spectral
resolution and extend.
}
\examples{

\dontrun{
data(shootout)
shootout_scans_x_2 <- combineSpectra(x = shootout_scans, y = shootout_scans)
}


}
\author{
Daniel M Griffith
}
\keyword{manipulation}
