% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSpectra.R
\name{convertSpectra}
\alias{convertSpectra}
\title{Converts spectra into new units.}
\usage{
convertSpectra(x, method = NULL)
}
\arguments{
\item{x}{An object of class \code{spectra.list} or \code{spectra.matrix}.}

\item{method}{The desired conversion. Selected from:
\cr "A_to_TR" - Absorbance to Transmittance.
\cr "TR_to_A" - Transmittance to Absorbance.
\cr "A_to_TR" - Absorbance to Transmittance.
\cr "TR_to_A" - Transmittance to Absorbance.
\cr "WN_to_WL" - Wavenumbers to Wavelength.
\cr "WL_to_WN" - Wavelength to Wavenumbers.}
}
\value{
A \code{spectra.matrix} containing the converted spectra.
}
\description{
This function accepts spectra in a single \code{spectra.list} or
\code{spectra.matrix} object performs a specified units conversion. 
Conversion between wavelengths and wave numbers are reliable.  However, 
conversions among transmittance, absorbance, and reflectance are
naive estimations (inverse log10 transformations) provided for convenience.
}
\examples{

\dontrun{
data(shootout)
attributes(shootout_scans)
converted_scans <- convertSpectra(x = shootout_scans, method = "WL_to_WN")
}


}
\author{
Daniel M Griffith
}
\keyword{manipulation}
