% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\name{extract-methods}
\alias{extract-methods}
\title{extract methods}
\arguments{
\item{x}{Object of class \code{spectra.matrix}.}

\item{...}{Additional args.}
}
\value{
Returns an object of class \code{spectra.matrix}.
}
\description{
extract methods
}
\author{
Daniel M Griffith
}
\keyword{manipulation}
