% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCompatible.R
\name{makeCompatible}
\alias{makeCompatible}
\title{Manipulate a spectral dataset to conform to the units, resolution, and
extent of a reference dataset.}
\usage{
makeCompatible(x, ref)
}
\arguments{
\item{x}{An object of class \code{spectra.list} or \code{spectra.matrix}
containing the spectra to modify.}

\item{ref}{An object of class \code{spectra.list} or \code{spectra.matrix}
have the attributes (units, resolution, etc.) desired for \code{x}.}
}
\value{
Returns an object of class \code{spectra.matrix} containing the modified spectra.
}
\description{
This function accepts spectra as a \code{spectra.list} or
\code{spectra.matrix} object, and manipulates these spectra to conform to the
units, resolution, and extent of a reference dataset. The transformations
occur in the following order:
 - Conversion between wavelength and wavenumber with \code{convertSpectra()}.
 - Conversion among absorbance, reflectance, transmittance values with \code{convertSpectra()}.
 - Resampling of spectra is done with \code{approx()}, which performs a linear interpolation.
}
\examples{

\dontrun{
data(shootout)
reference <- convertSpectra(x = shootout_scans, method = "WL_to_WN")
reference <- reference[,1:100] 
makeCompatible(x = shootout_scans, ref = reference)
}

}
\author{
Daniel M Griffith
}
\keyword{manipulation}
