% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspec-package.R
\docType{data}
\name{N_cal_shootout}
\alias{N_cal_shootout}
\title{N_cal_shootout: Shootout NIR calibration for Nitrogen}
\format{An object of class \code{PLScalibration}.}
\description{
NIR calibration for Nitrogen using the \code{shootout} data.
}
\examples{

}
\references{
McClure, W., 1998. Software shootout at the idrc98. In: The
Ninth International Diffuse Reflectance Conference, Chambersburg,
Pennsylvania.
}
\keyword{calibrations}
