% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspec-package.R
\docType{data}
\name{shootout}
\alias{shootout}
\alias{shootout_scans}
\alias{shootout_wetlab}
\title{Sample spectral data: software shootout at the IDRC98}
\format{shootout_scans - a \code{spectra.matrix} containing the sample NIR
spectra.  shootout_wetlab - a \code{data.frame} with the nitrogen and carbon
data.}
\description{
Sample NIR and wet lab data for illustration. Dry plant material (Fescue
grass) from a fertilizer experiment were ground in a Wiley mill and scanned
on a FOSS NIRSystems spectrophotometer in 1998. Carbon and nitrogen data
were determined independently with a LECO CNS-2000.
}
\examples{

data(shootout)

}
\references{
McClure, W., 1998. Software shootout at the idrc98. In: The
Ninth International Diffuse Reflectance Conference, Chambersburg,
Pennsylvania.
}
\keyword{datasets}
