% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSpectra.R
\name{writeSpectra}
\alias{writeSpectra}
\title{Function to write spectra.}
\usage{
writeSpectra(spectra, filelist = NULL, path = NULL)
}
\arguments{
\item{spectra}{An object of class \code{spectra.list} containing the spectra
to write.}

\item{filelist}{A vector of full file names, in the same order as the
elements of \code{spectra}.}

\item{path}{A string specifying a directory, where all spectra will be
written using \code{names(spectra)} as file names. Not used with
\code{filelist} - ignored if \code{filelist} is specified.}
}
\value{
Writes spectra to files.
}
\description{
This function accepts spectra in a single \code{spectra.list} or
\code{spectra.matrix} object and writes those spectra to separate files.
Currently this function only writes ".dpt" files.
}
\examples{

\dontrun{
data(shootout)
writeSpectra(spectra = shootout_scans, path = "MY_PATH")
}

}
\author{
Daniel M Griffith
}
\keyword{read/write/fileio}
